+++ README.md
@@ -6,4 +6,5 @@ WPFとUWPで使えるフレームワークを作ってみます。
 ⇒予定を変更して、当面以下の内容で進めます。
   1. .NET 5 ベース（.NET 6の予定）
   2. まずはログと国際化対応のみに絞る
+
 ※JaffaFramework フォルダ内は、最終的には廃棄する予定です。
