+++ docs/ja-jp/Refarence/Jaffa.Diagnostics.LogType.md
@@ -1,0 +1,21 @@
+# Jaffa.Diagnostics.LogType 列挙型
+
+LogType 列挙型は、ログメッセージの種類を表します。
+
+
+## 構文
+
+```
+public enum LogTypes : byte
+```
+
+## フィールド
+
+<table><tr><td>名前</td><td>説明</td></tr>
+
+<tr><td>Error</td><td>エラーメッセージです。</td></tr>
+<tr><td>Information</td><td>情報メッセージです。</td></tr>
+<tr><td>Warning</td><td>警告メッセージです。</td></tr>
+
+</table>
+
