+++ docs/ja-jp/Refarence/Jaffa.Diagnostics.md
@@ -1,0 +1,12 @@
+# Jaffa.Diagnostics 名前空間
+ 
+## クラス
+
+- [Logging クラス](Jaffa.Diagnostics.Logging.md)
+- [LoggingSettings クラス](Jaffa.Diagnostics.LoggingSettings.md)
+- [LogWritingEventArgs クラス](Jaffa.Diagnostics.LogWritingEventArgs.md)
+
+## 列挙型
+
+- [LoggingMode 列挙型](Jaffa.Diagnostics.LoggingMode.md)
+- [LogType 列挙型](Jaffa.Diagnostics.LogType.md)
