+++ src/JaffaCore/Diagnostics/LoggingMode.cs
@@ -12,19 +12,19 @@ namespace Jaffa.Diagnostics
         /// </summary>
         None,
         /// <summary>
-        /// 指定したサイズを超えたらファイルを切り替えます。ファイル名には(1-2)が入ります。常に1が新しいファイルです。
+        /// 指定したサイズを超えたらファイルを切り替えます。ファイル名には(01-02)が入ります。常に01が新しいファイルです。
         /// </summary>
         Size,
         /// <summary>
-        /// 日単位でファイルを切り替えます。ファイル名には日(1-31)が入ります。
+        /// 日単位でファイルを切り替えます。ファイル名には日(01-31)が入ります。
         /// </summary>
         Day,
         /// <summary>
-        /// 週単位でファイルを切り替えます。ファイル名には週(1-7)が入ります。
+        /// 週単位でファイルを切り替えます。ファイル名には週(01-07)が入ります。
         /// </summary>
         Week,
         /// <summary>
-        /// 月単位でファイルを切り替えます。ファイル名には月(1-12)が入ります。
+        /// 月単位でファイルを切り替えます。ファイル名には月(01-12)が入ります。
         /// </summary>
         Month
     }
